ANDROAPP-5373 Refactor DatasetTableUi and extract models to explicit classes (#3199)
