+++ compose-table/src/main/java/org/dhis2/composetable/actions/DefaultValidator.kt
@@ -1,0 +1,3 @@
+package org.dhis2.composetable.actions
+
+class DefaultValidator : Validator
