+++ compose-table/src/main/java/org/dhis2/composetable/actions/TableInteractions.kt
@@ -1,6 +1,5 @@
 package org.dhis2.composetable.actions
 
-import androidx.compose.runtime.compositionLocalOf
 import org.dhis2.composetable.model.TableCell
 import org.dhis2.composetable.model.TableDialogModel
 import org.dhis2.composetable.ui.TableSelection
@@ -11,5 +10,3 @@ interface TableInteractions {
     fun onClick(tableCell: TableCell) = run { }
     fun onOptionSelected(cell: TableCell, code: String, label: String) = run { }
 }
-
-val LocalInteraction = compositionLocalOf<TableInteractions> { object : TableInteractions {} }
