+++ compose-table/src/main/java/org/dhis2/composetable/actions/TableResizeActions.kt
@@ -1,7 +1,5 @@
 package org.dhis2.composetable.actions
 
-import androidx.compose.runtime.compositionLocalOf
-
 interface TableResizeActions {
     fun onTableWidthChanged(width: Int) = run {}
     fun onRowHeaderResize(tableId: String, newValue: Float) = run {}
@@ -9,6 +7,3 @@ interface TableResizeActions {
     fun onTableDimensionResize(tableId: String, newValue: Float) = run {}
     fun onTableDimensionReset(tableId: String) = run {}
 }
-
-val LocalTableResizeActions =
-    compositionLocalOf<TableResizeActions> { object : TableResizeActions {} }
