+++ compose-table/src/main/java/org/dhis2/composetable/actions/Validator.kt
@@ -1,6 +1,5 @@
 package org.dhis2.composetable.actions
 
-import androidx.compose.runtime.staticCompositionLocalOf
 import org.dhis2.composetable.model.TableCell
 import org.dhis2.composetable.model.ValidationResult
 
@@ -9,7 +8,3 @@ interface Validator {
         return ValidationResult.Success(tableCell.value)
     }
 }
-
-class DefaultValidator : Validator
-
-val LocalValidator = staticCompositionLocalOf<Validator> { DefaultValidator() }
