+++ compose-table/src/main/java/org/dhis2/composetable/model/HeaderMeasures.kt
@@ -1,0 +1,3 @@
+package org.dhis2.composetable.model
+
+data class HeaderMeasures(val width: Int, val height: Int)
