+++ compose-table/src/main/java/org/dhis2/composetable/model/TableCornerUiState.kt
@@ -1,0 +1,7 @@
+package org.dhis2.composetable.model
+
+data class TableCornerUiState(
+    val isSelected: Boolean = false,
+    val onTableResize: (Float) -> Unit,
+    val onResizing: (ResizingCell?) -> Unit
+)
