+++ compose-table/src/main/java/org/dhis2/composetable/model/TableHeaderRow.kt
@@ -1,0 +1,6 @@
+package org.dhis2.composetable.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TableHeaderRow(val cells: List<TableHeaderCell>)
