+++ compose-table/src/main/java/org/dhis2/composetable/model/extensions/TableModelExtensions.kt
@@ -1,0 +1,13 @@
+package org.dhis2.composetable.model.extensions
+
+import org.dhis2.composetable.model.TableModel
+
+fun TableModel.areAllValuesEmpty(): Boolean {
+    this.tableRows.forEach { it ->
+        val result = it.values.values.filterNot { it.value == "" }
+        if (result.isNotEmpty()) {
+            return false
+        }
+    }
+    return true
+}
