+++ compose-table/src/main/java/org/dhis2/composetable/ui/TableActions.kt
@@ -1,0 +1,43 @@
+package org.dhis2.composetable.ui
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.material.Icon
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import org.dhis2.composetable.R
+
+@Composable
+fun TableActions(modifier: Modifier, title: String, actionIcons: @Composable () -> Unit) {
+    Row(
+        modifier = modifier,
+        horizontalArrangement = Arrangement.Absolute.spacedBy(8.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Icon(
+            imageVector = ImageVector.vectorResource(id = R.drawable.ic_table),
+            contentDescription = "",
+            tint = TableTheme.colors.primary
+        )
+        Text(
+            modifier = Modifier.weight(1f),
+            text = title,
+            style = TextStyle(
+                color = Color.Black,
+                fontSize = 14.sp,
+                fontWeight = FontWeight.Normal,
+                lineHeight = 10.sp
+            )
+        )
+        actionIcons()
+    }
+}
