+++ compose-table/src/main/java/org/dhis2/composetable/ui/extensions/LazyListStateExtensions.kt
@@ -1,0 +1,8 @@
+package org.dhis2.composetable.ui.extensions
+
+import androidx.compose.foundation.gestures.animateScrollBy
+import androidx.compose.foundation.lazy.LazyListState
+
+suspend fun LazyListState.animateScrollToVisibleItems() {
+    animateScrollBy(layoutInfo.viewportSize.height / 2f)
+}
