+++ compose-table/src/main/java/org/dhis2/composetable/ui/modifiers/CornerBackground.kt
@@ -1,4 +1,4 @@
-package org.dhis2.composetable.ui
+package org.dhis2.composetable.ui.modifiers
 
 import androidx.compose.foundation.background
 import androidx.compose.ui.Modifier
